SPARKC-4: Add GettableDataToMappedTypeConverter.
GettableDataToMappedTypeConverter can create an object of class T from a GettableData instance, e.g. a CassandraRow or a UDTValue, using the column mapping provided by an appropriate ColumnMapper[T].
It is ok to nest UDTs inside of UDTs as well as inside of collections, because object fields and collection items are converted recursively.